requirements.txt
gradio
openai==0.28.1
pypdf
langchain
pymupdf
python-dotenv
tiktoken
chromadb==0.4.15